remove old full folder

+++ run_ANI.sh
@@ -191,7 +191,9 @@ if [[ ${sample_count} -gt ${max_ani_samples} ]]; then
 		echo "Moving ${filename}"
 		cp ${OUTDATADIR}/ANI/localANIDB/${filename} ${OUTDATADIR}/ANI/localANIDB_trimmed/
 	done
-	rm -r "${OUTDATADIR}/ANI/localANIDB_full"
+	if [[ -d "${OUTDATADIR}/ANI/localANIDB_full" ]]; then
+		rm -r "${OUTDATADIR}/ANI/localANIDB_full"
+	fi
 	mv "${OUTDATADIR}/ANI/localANIDB" "${OUTDATADIR}/ANI/localANIDB_full"
 	mv "${OUTDATADIR}/ANI/localANIDB_trimmed" "${OUTDATADIR}/ANI/localANIDB"
 # Continue without reducing the tree, as there are not enough samples to require reduction
